Remove the PipeRouter transient-retry loop
Direct (non-Temporal) execution is a single pipeline-level attempt on top of
a transport that already retries brief blips. The PipeRouter application-level
transient-retry loop reproduced Temporal's durability at the wrong layer: it
only ever ran on the direct path (dead code on the Temporal path), re-ran at
the wrong granularity, and snapshotted its retry budget from global config so
a per-run execution_config was silently ignored.

- pipe_router_protocol.py: run() is now _before_run -> one _run_pipe_job ->
  _after_successful_run; the except (CogtError, PipeRunError) handler is kept
  as error propagation (PipeRunError -> PipeRouterError with pipe-stack), not
  retry. Removed the while loop, backoff sleep, retry logging.
- Deleted transient_retry.py; dropped make_transient_retry_settings() and the
  transient_retry_settings attribute from PipeRouter / DryPipeRouter /
  TemporalPipeRouter.
- Removed the four transient_retry_* fields and the timing validator from
  PipelineExecutionConfig and all three pipelex.toml files. max_concurrency
  (bounded fan-out) stays.
- Replaced the retry tests with a guard test pinning both handler branches:
  direct = single attempt, and PipeRunError -> PipeRouterError.
- Fixed stale prose left false by the removal (find_inference_error_category_
  in_chain docstring, run_pipe stack-pop comment, Linkup CHANGELOG entry).
