+++ pipelex/pipe_run/pipe_router_protocol.py
@@ -1,21 +1,15 @@
-import asyncio
 from abc import abstractmethod
 from typing import Protocol
 
-from pipelex import log
-from pipelex.cogt.exceptions import CogtError, find_inference_error_category_in_chain
+from pipelex.cogt.exceptions import CogtError
 from pipelex.core.pipes.pipe_output import PipeOutput
 from pipelex.observer.observer_protocol import ObserverProtocol, PayloadKey, PayloadType
 from pipelex.pipe_run.exceptions import PipeRouterError, PipeRunError
 from pipelex.pipe_run.pipe_job import PipeJob
-from pipelex.pipe_run.transient_retry import TransientRetrySettings
 
 
 class PipeRouterProtocol(Protocol):
     observer: ObserverProtocol
-    # Resolved from `PipelineExecutionConfig` by each concrete router at construction time — the protocol
-    # itself cannot read config without forming an import cycle (see `transient_retry.py`).
-    transient_retry_settings: TransientRetrySettings
 
     async def _before_run(
         self,
@@ -57,43 +51,25 @@ class PipeRouterProtocol(Protocol):
     ) -> PipeOutput:
         await self._before_run(pipe_job)
 
-        retry_settings = self.transient_retry_settings
-        retry_count = 0
-        while True:
-            try:
-                pipe_output = await self._run_pipe_job(pipe_job)
-                break
-            except (CogtError, PipeRunError) as exc:
-                # The transient-retry decision is driven by the underlying inference error's
-                # category. The LLM operators (PipeLLM, PipeStructure) wrap the worker's
-                # CogtError into a PipeRunError before it reaches here, so the category is
-                # recovered by walking the full `__cause__` chain to any depth.
-                error_category = find_inference_error_category_in_chain(exc)
-                is_retryable = error_category is not None and error_category.is_retryable
-                if is_retryable and retry_count < retry_settings.max_transient_retries:
-                    retry_count += 1
-                    wait_seconds = retry_settings.compute_wait(retry_count)
-                    log.warning(
-                        f"Transient inference error ({error_category}) running pipe '{pipe_job.pipe.code}' — "
-                        f"retry {retry_count}/{retry_settings.max_transient_retries} in {wait_seconds:.2f}s: {exc}"
-                    )
-                    await asyncio.sleep(wait_seconds)
-                    continue
-                # Non-retryable category, or the transient-retry budget is exhausted. A PipeRunError
-                # still wraps into PipeRouterError (preserving the pipe location context); a raw
-                # CogtError is re-raised as-is so its cause chain is preserved.
-                await self._after_failing_run(pipe_job, exc)
-                if isinstance(exc, PipeRunError):
-                    raise PipeRouterError(
-                        message=exc.message,
-                        run_mode=pipe_job.pipe_run_params.run_mode,
-                        pipe_code=pipe_job.pipe.code,
-                        output_name=pipe_job.output_name,
-                        # run_pipe() has already popped the failed pipe's own frame; re-append
-                        # its code so the reported stack still ends with the pipe that failed.
-                        pipe_stack=[*pipe_job.pipe_run_params.pipe_stack, pipe_job.pipe.code],
-                    ) from exc
-                raise
+        try:
+            pipe_output = await self._run_pipe_job(pipe_job)
+        except (CogtError, PipeRunError) as exc:
+            # Direct (non-Temporal) execution is a single pipeline-level attempt — there is no
+            # retry here. This handler is error propagation, not retry: a PipeRunError wraps into
+            # PipeRouterError (preserving the pipe location context); a raw CogtError is re-raised
+            # as-is so its cause chain is preserved. Resilience is the Temporal track's job.
+            await self._after_failing_run(pipe_job, exc)
+            if isinstance(exc, PipeRunError):
+                raise PipeRouterError(
+                    message=exc.message,
+                    run_mode=pipe_job.pipe_run_params.run_mode,
+                    pipe_code=pipe_job.pipe.code,
+                    output_name=pipe_job.output_name,
+                    # run_pipe() has already popped the failed pipe's own frame; re-append
+                    # its code so the reported stack still ends with the pipe that failed.
+                    pipe_stack=[*pipe_job.pipe_run_params.pipe_stack, pipe_job.pipe.code],
+                ) from exc
+            raise
 
         await self._after_successful_run(pipe_job, pipe_output)
