+++ pipelex/pipelex.toml
@@ -286,12 +286,6 @@ is_normalize_data_urls_to_storage = true
 is_mock_inputs = false
 is_generate_graph = true
 
-# Retry of transient inference failures (rate limits, timeouts, brief outages) at the dispatch layer.
-max_transient_retries = 3                # 0 disables retry entirely; the default is a small sane budget
-transient_retry_base_wait = 2.0          # seconds before the first retry
-transient_retry_max_wait = 30.0          # cap on the per-retry wait
-transient_retry_backoff_multiplier = 2.0 # the wait is multiplied by this on each successive retry
-
 # Bounded fan-out concurrency for PipeBatch: at most this many branches run at once.
 # A modest default keeps a large workload (one pipe over many items) from overwhelming
 # asyncio, memory, and provider rate limits. Set to "unbounded" to disable the bound.
