+++ wip/error-handling/README.md
@@ -11,7 +11,7 @@ The improvement plan lives elsewhere; the docs here describe the ground that any
 | Error metadata model | Landed for all inference workers — inference errors self-describe via `to_error_report()`, and every LLM, img-gen, extract, and search worker raise carries structured `ProviderErrorMetadata` (status_code, request_id, retry_after_seconds, provider_error_code, body) plus `UserAction(kind, detail)` with semantic `UserActionKind` values. `InferenceErrorCategory.UNKNOWN` distinguishes unrecognized fallback paths from genuine `CONTENT`. Each worker family also fills `model` / `provider` onto the error at its public-method chokepoint via `CogtError.fill_model_and_provider()`, so a production inference failure's `ErrorReport` is attributable to a model and a backend — previously both came back `None` outside of tests. Non-inference paths still depend on string-keyed dicts in `agent_output.py`. | [track-metadata-model.md](track-metadata-model.md) |
 | Worker classification | Landed for all inference workers — every LLM, img-gen, extract, and search worker (plus AWS Bedrock LLM) maps SDK exceptions to a `CogtError` subclass with `error_category`, unwraps `InstructorRetryException` correctly on structured-gen paths, and carries the structured metadata + semantic user-action upgrades. Response-shape validation failures (non-SDK paths) carry `provider_metadata=None` plus a semantic `UserAction` for uniformity. | [track-worker-classification.md](track-worker-classification.md) |
 | Extract / Classify / Render | Proposed (not started) — decomposes the per-worker pipeline into one per-provider Extract function + shared Classify + shared Render. Cuts duplication across 18+ workers. The worker-classification sweep it depended on has landed, so it is now unblocked. | [track-extract-classify-render.md](track-extract-classify-render.md) |
-| Retry & resilience | **Superseded — see [track-retry-and-resilience.md](track-retry-and-resilience.md): the target architecture removes the application-level retry loop (direct execution = single attempt; resilience = Temporal).** Landed in current code — `PipeRouter` retries `TRANSIENT` `CogtError`s with config-driven exponential backoff (`max_transient_retries`, default 3); `PipeBatch` bounds fan-out via `gather_bounded` (chunked, `max_concurrency`, default 8). `tenacity` retry removed from the gateway workers. The router's retry decision unifies the raw-`CogtError` and `PipeRunError`-wrapping-a-`CogtError` cases, so transient LLM failures (wrapped by `PipeLLM` / `PipeStructure`) are retried in-process too — see the archived [todos-llm-retry-loop-bypass.md](todos-llm-retry-loop-bypass.md). | [track-retry-and-resilience.md](track-retry-and-resilience.md) |
+| Retry & resilience | The application-level `PipeRouter` transient-retry loop has been removed — direct (non-Temporal) execution is a single pipeline-level attempt, and resilience is the Temporal track's job (activity `RetryPolicy` keyed off `InferenceErrorCategory`). `PipeBatch` still bounds fan-out via `gather_bounded` (chunked, `max_concurrency`, default 8) — admission control, not retry. Tier 1 (transport retry) is the provider SDKs' own retry of brief blips; making it an explicit, uniform policy is the open Workstream 2 in the track doc. | [track-retry-and-resilience.md](track-retry-and-resilience.md) |
 | CLI delivery | Landed — agent CLI emits markdown by default for `run`/`validate`/`init` (and `models`/`check-model`/`doctor`), `--format json` for the structured payload; the error path follows the same option via a per-invocation `ContextVar`. `ErrorReport` carries an authoritative `error_domain` → HTTP-status mapping. `display_error_panel` collapses the field-shaped Rich handlers onto one helper. | [track-cli-delivery.md](track-cli-delivery.md) |
 | Temporal integration | Landed — `TemporalError.from_message_exception()` derives retryability from `InferenceErrorCategory.is_retryable` for category-carrying `CogtError`s (the `non_retryable_error_types` list is the fallback) and packs the full `ErrorReport` into `ApplicationError.details`. Every in-scope activity is decorated with `@convert_pipelex_errors`, so the bridge now runs in production — see Followup 5 in the track doc. | [track-temporal-integration.md](track-temporal-integration.md) |
 | Testing | Landed — worker-level classification tests, the `agent_output.py` dict-drift test, and the full-chain worker→runner→CLI integration test (JSON + markdown) plus Rich-panel snapshot tests are all in place. | [track-testing.md](track-testing.md) |
